% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_CINE.R
\docType{data}
\name{df_CINE}
\alias{df_CINE}
\title{Peru: International Standard Classification of Education}
\format{
dataframe
\describe{
  \item{Education_COD}{chr Education code}
  \item{Education}{chr Education}
  \item{Specific_COD}{chr Specific code}
  \item{Specific}{chr Specific}
  \item{AcademicLevel}{dbl Academic Level}
  \item{AcademicLevel_COD}{dbl Academic Level code}
  \item{EducationProgram}{dbl Education Program}
  \item{EducationProgram_Lemma}{dbl Lemma of the Education Program}
}
}
\usage{
df_CINE
}
\description{
Peru: International Standard Classification of Education
}
\keyword{datasets}
