% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_LEMMA.R
\docType{data}
\name{df_LEMMA}
\alias{df_LEMMA}
\title{Lemma of the Education Program}
\format{
dataframe
\describe{
  \item{word}{chr word}
  \item{lemma}{chr lemma}
}
}
\usage{
df_LEMMA
}
\description{
Lemma of the Education Program
}
\keyword{datasets}
